refactor: re-structure repo
