+++ .env.sample
@@ -11,6 +11,8 @@ MYSQL_USER=
 MYSQL_PASSWORD=
 
 # Send mail
+# Ensure enable 2-step verification
+# https://myaccount.google.com/apppasswords
 MAIL_HOST=
 MAIL_PASS=
