+++ pom.xml
@@ -40,7 +40,7 @@
                         <configuration>
                             <archive>
                                 <manifest>
-                                    <mainClass>runtime.Main</mainClass>
+                                    <mainClass>Main</mainClass>
                                 </manifest>
                             </archive>
                         </configuration>
@@ -59,7 +59,7 @@
                             </execution>
                         </executions>
                         <configuration>
-                            <mainClass>runtime.Main</mainClass>
+                            <mainClass>Main</mainClass>
                             <classpathScope>test</classpathScope>
                         </configuration>
                     </plugin>
