+++ src/main/java/constants/Info.java
@@ -1,6 +1,6 @@
 package constants;
 
-import utils.UIUtils;
+import views.UIPrompts;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
@@ -9,7 +9,7 @@ public class Info implements ActionListener {
 
     @Override
     public void actionPerformed(ActionEvent e) {
-        UIUtils.IS_ABOUT_ME();
+        UIPrompts.IS_ABOUT_ME();
     }
 
 }
