+++ src/main/java/controllers/ToggleHandler.java
@@ -1,4 +1,4 @@
-package utils;
+package controllers;
 
 import views.Toggle;
