+++ src/main/java/models/ui/RegisterModel.java
@@ -2,8 +2,8 @@ package models.ui;
 
 import constants.Regex;
 import errors.DataException;
-import models.data.Account;
-import services.UserDAO;
+import modules.user.UserDTO;
+import modules.user.UserDAO;
 import utils.PasswordHandler;
 
 public record RegisterModel(String username, String password, String confirmPassword) {
@@ -36,7 +36,7 @@ public record RegisterModel(String username, String password, String confirmPass
     }
 
     public boolean isDuplicateUsername(String username) {
-        Account db;
+        UserDTO db;
         UserDAO executeQuery = UserDAO.getInstance();
         try {
             db = executeQuery.selectEmailAndPasswordByEmail(username);
