--- src/main/java/module-info.java
@@ -1,7 +1,0 @@
-module Java.Snake.Game {
-    requires java.desktop;
-    requires java.sql;
-    requires io.github.cdimascio.dotenv.java;
-    requires jdk.httpserver;
-    requires javax.mail.api;
-}
