+++ src/main/java/modules/board/Mill.java
@@ -1,4 +1,4 @@
-package models.board;
+package modules.board;
 
 import constants.ResourcePaths;
 import styles.UISizes;
@@ -7,16 +7,22 @@ import javax.swing.*;
 import java.awt.*;
 import java.io.InputStream;
 
-public class Box extends Board {
+public class Mill extends Board {
     protected int wallThickness = 20;
     private Image wall;
 
-    public Box() {
+    public  Mill() {
         super();
+
+        //adjust to prevent hitting the side wall
+        leftDirection = false;
+        rightDirection = false;
+        upDirection = false;
+        downDirection = true;
     }
 
     @Override
-    protected void loadImages() {
+    public void loadImages() {
         super.loadImages();
         wall = new ImageIcon(getClass().getResource(ResourcePaths.URL_WALL)).getImage();
     }
@@ -24,14 +30,23 @@ public class Box extends Board {
     @Override
     public void paintComponent(Graphics g) {
         super.paintComponent(g);
-        for (int i = 0; i < UISizes.WIDTH_BOARD; i += 20) {
-            g.drawImage(wall, i, 0, this);
-            g.drawImage(wall, i, UISizes.HEIGHT_BOARD - 70, this);
+        //draw the left horizontal wall from x = 0 to x = 140 at y = 420
+        for (int i = 0; i < 160; i += 20) {
+            g.drawImage(wall, i, UISizes.HEIGHT_BOARD - 50 - 80, this);
         }
-        for (int i = 0; i < UISizes.HEIGHT_BOARD - 70; i += 20) {
-            g.drawImage(wall, 0, i, this);
-            g.drawImage(wall, UISizes.WIDTH_BOARD - 20, i, this);
+        //draw the right horizontal wall from x = 340 to x = 500 at y = 60
+        for (int i = 340; i < UISizes.WIDTH_BOARD; i += 20) {
+            g.drawImage(wall, i, 60, this);
         }
+        //draw the vertical walls at x = 140, from y = 0 to y = 340
+        for (int i = 0; i < 340; i += 20) {
+            g.drawImage(wall, 140, i, this);
+        }
+        //draw the vertical walls at x = 340, from y = 500 to y = 160
+        for (int i = 160; i < 500; i += 20) {
+            g.drawImage(wall, 340, i, this);
+        }
+
     }
 
     @Override
@@ -44,19 +59,43 @@ public class Box extends Board {
             }
         }
 
-        if (y[0] >= UISizes.HEIGHT_BOARD - 70) {
-            inGame = false;
+        // snake can go through the walls north
+        if (y[0] < 0) {
+            y[0] = UISizes.HEIGHT_BOARD - 50 - DOT_SIZE;
         }
 
-        if (y[0] < 20) {
-            inGame = false;
+        // snake can go through the walls south
+        if(y[0] >= UISizes.HEIGHT_BOARD - 50){
+            y[0] = 0;
         }
 
+        // snake can go through the walls east
         if (x[0] >= UISizes.WIDTH_BOARD) {
+            x[0] = 0;
+        }
+
+        // snake can go through the walls west
+        if (x[0] < 0) {
+            x[0] = UISizes.WIDTH_BOARD - DOT_SIZE;
+        }
+
+        // snake hit the vertical wall at x = 140, from y = 0 to y = 340
+        if(x[0] == 140 && y[0] < 340 && y[0] > 0){
             inGame = false;
         }
 
-        if (x[0] < 20) {
+        // snake hit the vertical wall at x = 340, from y = 500 to y = 160
+        if(x[0] == 340 && y[0] < 500 && y[0] > 160){
+            inGame = false;
+        }
+
+        // snake hit the horizontal wall from x = 0 to x = 140 at y = 420
+        if(y[0] == UISizes.HEIGHT_BOARD - 50 - 80 && x[0] < 140 && x[0] > 0){
+            inGame = false;
+        }
+
+        // snake hit the horizontal wall from x = 340 to x = 500 at y = 60
+        if(y[0] == 60 && x[0] < UISizes.WIDTH_BOARD && x[0] > 340){
             inGame = false;
         }
