+++ src/main/java/modules/otp/OTPUtils.java
@@ -1,4 +1,4 @@
-package utils;
+package modules.otp;
 
 import javax.swing.JOptionPane;
 
@@ -8,6 +8,6 @@ public class OTPUtils {
     };
 
     public static void IS_NOTIFY_VERIFY_ACCOUNT() {
-        JOptionPane.showMessageDialog(null, "We've send your otp back to your email, please type back to authenticate!", "Snake Game", JOptionPane.INFORMATION_MESSAGE);
+        JOptionPane.showMessageDialog(null, "We've send your otp back to your email, please provide back to authenticate!", "Snake Game", JOptionPane.INFORMATION_MESSAGE);
     }
 }
