+++ src/main/java/modules/phone/PhoneUtils.java
@@ -1,4 +1,4 @@
-package utils;
+package modules.phone;
 
 public class PhoneUtils {
