+++ src/main/java/modules/sound/AudioHandler.java
@@ -1,4 +1,4 @@
-package utils;
+package modules.sound;
 
 import errors.FileException;
