+++ src/test/java/views/RegisterViewTest.java
@@ -1,6 +1,6 @@
 package views;
 
-import models.data.Account;
+import modules.user.UserDTO;
 import org.junit.Before;
 import org.junit.Test;
 import utils.EnvUtils;
@@ -26,14 +26,15 @@ public class RegisterViewTest {
 
     @Test
     public void getRegister() {
-        Account expectedResult = new Account(email, firstName, lastName,password, password);
+        UserDTO expectedResult = new UserDTO(email, firstName, lastName,password, password);
 
         //reg_date above and below is different because the constructor of Account class has a reg_date parameter
         //depend on the current time, so we can't compare the whole object
         registerView.setRegister(email, firstName, lastName, password, password);
-        Account actualAccount = registerView.getRegister();
+        UserDTO actualUserDTO = registerView.getRegister();
 
-        boolean actualResult = expectedResult.email().equals(actualAccount.email()) && expectedResult.password().equals(actualAccount.password());
+        boolean actualResult = expectedResult.email().equals(actualUserDTO.email()) && expectedResult.password().equals(
+            actualUserDTO.password());
         assertTrue(actualResult);
     }
 }
